merged to main

+++ src/components/EditCategory.svelte
@@ -158,9 +158,6 @@
     }
 
     onMount(async () => { getCategory() })
-
-    $: console.log(category)
-    $: console.log(editCategory)
 </script>
 
 {#if !loading}
@@ -213,7 +210,7 @@
                         <p class="text-xs text-gray-400">{$LL.EditCategories.CategoryTitleDesc()}</p>
                         {#each editCategory.translations as translation}
                             {#if translation.language === languageShow}
-                                <input bind:value={translation.title} class="border my-1 p-2 peer rounded text-xs w-auto text-black" maxlength="75" name="titleCategory" />
+                                <input bind:value={translation.title} class="border my-1 p-2 peer rounded text-xs w-auto text-black" maxlength="100" name="titleCategory" />
                             {/if}
                         {/each}
                     </div>
@@ -222,7 +219,7 @@
                         <p class="text-xs text-gray-400">{$LL.EditCategories.CategoryDescDesc()}</p>
                         {#each editCategory.translations as translation}
                             {#if translation.language === languageShow}
-                                <textarea bind:value={translation.description} class="border my-1 p-2 peer rounded text-xs w-auto text-black" maxlength="250" name="descCategory" rows="4" />
+                                <textarea bind:value={translation.description} class="border my-1 p-2 peer rounded text-xs w-auto text-black" maxlength="300" name="descCategory" rows="4" />
                             {/if}
                         {/each}
                     </div>
@@ -251,16 +248,18 @@
                                                 <div class="flex flex-col flex-grow">
                                                     <label class="flex flex-col font-medium rounded text-xs w-full text-gray-900">
                                                         {$LL.EditCategories.QuestionTitle()}
-                                                        <input bind:value={rtoTranslation.title} class="border font-normal px-2 py-1 rounded text-sm border-gray-300" maxlength="75" type="text" />
+                                                        <input bind:value={rtoTranslation.title} class="border font-normal px-2 py-1 rounded text-sm border-gray-300" maxlength="100" type="text" />
                                                     </label>
                                                     <label class="flex flex-col font-medium rounded text-xs w-full text-gray-900">
                                                         {$LL.EditCategories.QuestionDesc()}
-                                                        <textarea bind:value={rtoTranslation.description} class="border font-normal px-2 py-1 rounded text-sm border-gray-300" maxlength="250" rows="2" />
+                                                        <textarea bind:value={rtoTranslation.description} class="border font-normal px-2 py-1 rounded text-sm border-gray-300" maxlength="300" rows="2" />
                                                     </label>
                                                 </div>
-                                                <button on:click={() => deleteQuestion(index)} class="hover:text-red-500 text-gray-400" title={$LL.EditCategories.DeleteQuestion()}>
-                                                    <svelte:component this={Trash2Icon} size={20} />
-                                                </button>
+                                                {#if !category.isBeingUsed || (!question.questionId && category.isBeingUsed)}
+                                                    <button on:click={() => deleteQuestion(index)} class="hover:text-red-500 text-gray-400" title={$LL.EditCategories.DeleteQuestion()}>
+                                                        <svelte:component this={Trash2Icon} size={20} />
+                                                    </button>
+                                                {/if}
                                             {:else}
                                                 <div class="flex flex-col flex-grow">
                                                     <p class="font-medium text-base text-gray-900">{rtoTranslation.title ? rtoTranslation.title : $LL.EditCategories.NoTitle()}</p>
@@ -302,19 +301,21 @@
                                                 <div class="flex flex-col flex-grow">
                                                     <label class="flex flex-col font-medium rounded text-xs w-full text-gray-900">
                                                         {$LL.EditCategories.QuestionTitle()}
-                                                        <input bind:value={rtoTranslation.title} class="border font-normal px-2 py-1 rounded text-sm border-gray-300" maxlength="75" type="text" />
+                                                        <input bind:value={rtoTranslation.title} class="border font-normal px-2 py-1 rounded text-sm border-gray-300" maxlength="100" type="text" />
                                                     </label>
                                                     <label class="flex flex-col font-medium rounded text-xs w-full text-gray-900">
                                                         {$LL.EditCategories.QuestionDesc()}
-                                                        <textarea bind:value={rtoTranslation.description} class="border font-normal px-2 py-1 rounded text-sm border-gray-300" maxlength="250" rows="2" />
+                                                        <textarea bind:value={rtoTranslation.description} class="border font-normal px-2 py-1 rounded text-sm border-gray-300" maxlength="300" rows="2" />
                                                     </label>
                                                 </div>
-                                                <button on:click={() => deleteQuestion(index)} class="hover:text-red-500 text-gray-400" title={$LL.EditCategories.DeleteQuestion()}>
-                                                    <svelte:component this={Trash2Icon} size={20} />
-                                                </button>
+                                                {#if !category.isBeingUsed || (!category.categoryId && category.isBeingUsed)}
+                                                    <button on:click={() => deleteQuestion(index)} class="hover:text-red-500 text-gray-400" title={$LL.EditCategories.DeleteQuestion()}>
+                                                        <svelte:component this={Trash2Icon} size={20} />
+                                                    </button>
+                                                {/if}
                                             {:else}
                                                 <div class="flex flex-col flex-grow">
-                                                    <p class="font-medium text-base text-gray-900">{rtoTranslation.title}</p>
+                                                    <p class="font-medium text-base text-gray-900">{rtoTranslation.title ? rtoTranslation.title : $LL.EditCategories.NoTitle()}</p>
                                                     <p class="text-xs text-gray-400">{rtoTranslation.description}</p>
                                                 </div>
                                             {/if}
